// types/transcation/Transaction.ts
export type TransactionKey =
    | "issue"
    | "receipt"
    | "purchase"
    | "purchase_return";
// ...
export interface PurchasePayload {


    ITEMID: number | null;
    
    SNO?:string;

    PCS: number;

    GRSWT: number;
    STNWT: number;
    NETWT: number;

    WASTYPE:string;
    // WASPER:number;
    // WASTAGE:number;

    
    TOUCH: number;
    PUREWT: number;

    HMC:number;

    STNAMT:number;
    MC:number;
    // ATOUCH:number;
    DESCRIPTION?:string;
}